requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai==1.3.0
tiktoken==0.5.1
pytesseract==0.3.10
pillow==10.1.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
python-dotenv==1.0.0
requests==2.31.0
googletrans==4.0.0rc1
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1
celery==5.3.4
pydantic==2.5.0
PyMuPDF==1.23.8
opencv-python==4.8.1.78
numpy==1.24.3
langdetect==1.0.9
langdetect googletrans==4.0.0rc1